app/requirements.txt
Flask
transbank-sdk
